Declaration int -> size_t for the VM (again)

+++ code/qcommon/vm_local.h
@@ -135,7 +135,7 @@ struct vm_s {
 
 	// for dynamic linked modules
 	void		*dllHandle;
-	int			(QDECL *entryPoint)( int callNum, ... );
+	int			(QDECL *entryPoint)( size_t callNum, ... );
 
 	// for interpreted modules
 	qboolean	currentlyInterpreting;
